+++ VComponentsDemo/Scenes/Demos/Modals/VModalDemoView.swift
@@ -16,12 +16,12 @@ struct VModalDemoView: View {
     @State private var isPresented: Bool = false
     @State private var hasTitle: Bool = true
     @State private var hasDivider: Bool = VModalModel.Layout().headerDividerHeight > 0
-    @State private var dismissType: Set<VModalModel.Misc.DismissType> = .default
+    @State private var dismissType: VModalModel.Misc.DismissType = .default
     
     private var model: VModalModel {
         var model: VModalModel = .init()
         
-        model.layout.headerDividerHeight = hasDivider ? (model.layout.headerDividerHeight == 0 ? 1 : model.layout.headerDividerHeight) : 0
+        model.layout.headerDividerHeight = hasDivider ? (model.layout.headerDividerHeight == 0 ? 1/3 : model.layout.headerDividerHeight) : 0
         model.colors.headerDivider = hasDivider ? (model.colors.headerDivider == .clear ? .gray : model.colors.headerDivider) : .clear
         
         model.misc.dismissType = dismissType
@@ -36,7 +36,10 @@ struct VModalDemoView: View {
     }
     
     private func component() -> some View {
-        VPlainButton(action: { isPresented = true }, title: "Present")
+        VPlainButton(
+            action: { isPresented = true },
+            title: "Present"
+        )
             .if(hasTitle,
                 ifTransform: {
                     $0
@@ -67,7 +70,7 @@ struct VModalDemoView: View {
                 .frame(maxWidth: .infinity, alignment: .leading)
             
             HStack(content: {
-                ForEach(VModalModel.Misc.DismissType.allCases, id: \.rawValue, content: { position in
+                ForEach(VModalModel.Misc.DismissType.all.elements(), id: \.rawValue, content: { position in
                     dismissTypeView(position)
                 })
             })
@@ -95,17 +98,29 @@ struct VModalDemoView: View {
     private var modalContent: some View {
         ZStack(content: {
             ColorBook.accent
-            
+
             if dismissType.isEmpty {
                 VStack(content: {
                     VText(
                         type: .multiLine(alignment: .center, limit: nil),
-                        color: ColorBook.primary,
+                        color: ColorBook.primaryWhite,
                         font: .system(size: 14, weight: .semibold),
                         title: "When close button is \"none\", Modal can only be dismissed programatically"
                     )
-                    
-                    VPlainButton(action: { isPresented = false }, title: "Dismiss")
+
+                    VPlainButton(
+                        model: {
+                            var model: VPlainButtonModel = .init()
+                            model.colors.title = .init(
+                                enabled: ColorBook.primaryWhite,
+                                pressed: ColorBook.primaryWhite,
+                                disabled: ColorBook.primaryWhite
+                            )
+                            return model
+                        }(),
+                        action: { isPresented = false },
+                        title: "Dismiss"
+                    )
                 })
                     .frame(maxHeight: .infinity, alignment: .bottom)
                     .padding(15)
@@ -121,7 +136,7 @@ extension VModalModel.Misc.DismissType {
         case .leadingButton: return "Leading"
         case .trailingButton: return "Trailing"
         case .backTap: return "Back Tap"
-        @unknown default: fatalError()
+        default: fatalError()
         }
     }
 }
@@ -130,5 +145,30 @@ extension VModalModel.Misc.DismissType {
 struct VModalDemoView_Previews: PreviewProvider {
     static var previews: some View {
         VModalDemoView()
+    }
+}
+
+
+
+public extension OptionSet where RawValue: FixedWidthInteger { // ???
+    func _elements() -> AnySequence<Self> {
+        var remainingBits = rawValue
+        var bitMask: RawValue = 1
+        return AnySequence {
+            return AnyIterator {
+                while remainingBits != 0 {
+                    defer { bitMask = bitMask &* 2 }
+                    if remainingBits & bitMask != 0 {
+                        remainingBits = remainingBits & ~bitMask
+                        return Self(rawValue: bitMask)
+                    }
+                }
+                return nil
+            }
+        }
+    }
+    
+    func elements() -> [Self] {
+        .init(_elements())
     }
 }