+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - `VBaseView` has been removed
 - `StateColors`'s and `StateOpacities`'s have been replaced with`GenericStateModel`'s
 - `VMenuPicker`/`Menu` label issue with iOS `15` has been fixed
+- Modal component sizes have been fixed in landscape orientation
 - `VSquareButton` font has been changed to system bold size 15
 - `VPlainButton` font has been changed to system medium size 15
 - `VWebLink` has been renamed to `VLink`